Use MethodHandles instead of Method invocations + Make FormattedCommandAliasProvider an interface
