+++ providers/BaseProviders/src/main/java/net/ess3/provider/FormattedCommandAliasProvider.java
@@ -3,28 +3,13 @@ package net.ess3.provider;
 import org.bukkit.command.CommandSender;
 import org.bukkit.command.FormattedCommandAlias;
 
-import java.util.ArrayList;
 import java.util.List;
 
-public abstract class FormattedCommandAliasProvider {
+public interface FormattedCommandAliasProvider extends Provider {
 
-    public List<String> createCommands(FormattedCommandAlias command, CommandSender sender, String[] args) {
-        final List<String> commands = new ArrayList<>();
-        for (String formatString : getFormatStrings(command)) {
-            final String cmd;
-            try {
-                cmd = buildCommand(command, sender, formatString, args);
-            } catch (Throwable th) {
-                continue; // Ignore, let server handle this.
-            }
+    List<String> createCommands(FormattedCommandAlias command, CommandSender sender, String[] args);
 
-            if (cmd == null) continue;
-            commands.add(cmd.trim());
-        }
-        return commands;
-    }
+    String[] getFormatStrings(FormattedCommandAlias command);
 
-    public abstract String[] getFormatStrings(FormattedCommandAlias command);
-
-    public abstract String buildCommand(FormattedCommandAlias command, CommandSender sender, String formatString, String[] args);
+    String buildCommand(FormattedCommandAlias command, CommandSender sender, String formatString, String[] args);
 }
