+++ providers/NMSReflectionProvider/src/main/java/net/ess3/nms/refl/providers/ReflFormattedCommandAliasProvider.java
@@ -5,35 +5,62 @@ import net.ess3.provider.FormattedCommandAliasProvider;
 import org.bukkit.command.CommandSender;
 import org.bukkit.command.FormattedCommandAlias;
 
+import java.lang.invoke.MethodHandle;
+import java.lang.invoke.MethodHandles;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
+import java.util.ArrayList;
+import java.util.List;
 
-public class ReflFormattedCommandAliasProvider extends FormattedCommandAliasProvider {
+public class ReflFormattedCommandAliasProvider implements FormattedCommandAliasProvider {
 
     private final boolean paper;
     private final Field formatStringsField;
-    private final Method buildCommandMethod;
+    private final MethodHandle buildCommandMethodHandle;
 
     public ReflFormattedCommandAliasProvider(boolean paper) {
         this.paper = paper;
 
         final Class<? extends FormattedCommandAlias> formattedCommandAliasClass;
         Field formatStringsField = null;
-        Method buildCommandMethod = null;
+        MethodHandle buildCommandMethodHandle = null;
         try {
             formattedCommandAliasClass = FormattedCommandAlias.class;
             formatStringsField = ReflUtil.getFieldCached(formattedCommandAliasClass, "formatStrings");
+
+            final Class<?>[] parameterTypes;
             if (paper) {
-                buildCommandMethod = ReflUtil.getMethodCached(formattedCommandAliasClass, "buildCommand", CommandSender.class, String.class, String[].class);
+                parameterTypes = new Class[] {CommandSender.class, String.class, String[].class};
             } else {
-                buildCommandMethod = ReflUtil.getMethodCached(formattedCommandAliasClass, "buildCommand", String.class, String[].class);
+                parameterTypes = new Class[] {String.class, String[].class};
             }
+
+            final Method buildCommandMethod = ReflUtil.getMethodCached(formattedCommandAliasClass, "buildCommand", parameterTypes);
+            buildCommandMethod.setAccessible(true);
+            buildCommandMethodHandle = MethodHandles.lookup().unreflect(buildCommandMethod);
         } catch (final Exception ex) {
             ex.printStackTrace();
         } finally {
             this.formatStringsField = formatStringsField;
-            this.buildCommandMethod = buildCommandMethod;
+            this.buildCommandMethodHandle = buildCommandMethodHandle;
         }
+    }
+
+    @Override
+    public List<String> createCommands(FormattedCommandAlias command, CommandSender sender, String[] args) {
+        final List<String> commands = new ArrayList<>();
+        for (String formatString : getFormatStrings(command)) {
+            final String cmd;
+            try {
+                cmd = buildCommand(command, sender, formatString, args);
+            } catch (Throwable th) {
+                continue; // Ignore, let server handle this.
+            }
+
+            if (cmd == null) continue;
+            commands.add(cmd.trim());
+        }
+        return commands;
     }
 
     @Override
@@ -49,12 +76,17 @@ public class ReflFormattedCommandAliasProvider extends FormattedCommandAliasProv
     public String buildCommand(FormattedCommandAlias command, CommandSender sender, String formatString, String[] args) {
         try {
             if (paper) {
-                return (String) buildCommandMethod.invoke(command, sender, formatString, args);
+                return (String) buildCommandMethodHandle.invoke(command, sender, formatString, args);
             } else {
-                return (String) buildCommandMethod.invoke(command, formatString, args);
+                return (String) buildCommandMethodHandle.invoke(command, formatString, args);
             }
-        } catch (ReflectiveOperationException ex) {
+        } catch (Throwable ex) {
             throw new RuntimeException(ex); // If this happens we have bigger problems...
         }
+    }
+
+    @Override
+    public String getDescription() {
+        return "NMS Reflection Provider for FormattedCommandAlias methods";
     }
 }
